Fix -Wshadow warnings by making OptimizationMarker an enum class
Bug: v8:12244,v8:12245
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3266004
Commit-Queue: Zhi An Ng <zhin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#77781}

+++ src/builtins/arm/builtins-arm.cc
@@ -892,7 +892,7 @@ static void TailCallRuntimeIfMarkerEquals(MacroAssembler* masm,
                                           Runtime::FunctionId function_id) {
   ASM_CODE_COMMENT(masm);
   Label no_match;
-  __ cmp_raw_immediate(actual_marker, expected_marker);
+  __ cmp_raw_immediate(actual_marker, static_cast<int>(expected_marker));
   __ b(ne, &no_match);
   GenerateTailCallToReturnedCode(masm, function_id);
   __ bind(&no_match);
